# Add DataReader, AvroDataReader, and integrate with GAME drivers