This is an incremental step towards the new client API. This change introduces a new interface -- **DataReader**, whose concrete implementations are responsible for reading data files of various formats and producing a DataFrame (and optionally an IndexMap). 

| DataReader |
| --- |
| `readMerged(path: String, featureColumnMap: Map[String, Set[String]]): (DataFrame, Map[String, IndexMapLoader])` |
| `readMerged(path: String, indexMapLoaders: Map[String, IndexMapLoader], featureColumnMap: Map[String, Set[String]]): DataFrame` |
| `read(path: String): (DataFrame, IndexMapLoader)` |
| `read(path: String, indexMapLoaders: Map[String, IndexMapLoader]): DataFrame` |
| `readMerged(paths: Seq[String], featureColumnMap: Map[String, Set[String]]): (DataFrame, Map[String, IndexMapLoader])` |
| `readMerged(paths: Seq[String], indexMapLoaders: Map[String, IndexMapLoader], featureColumnMap: Map[String, Set[String]]): DataFrame` |
| `read(paths: Seq[String]): (DataFrame, IndexMapLoader)` |
| `read(paths: Seq[String], indexMapLoaders: Map[String, IndexMapLoader]): DataFrame` |

The first such implementation is **AvroDataReader**, which has been integrated here into the GAME training and scoring drivers to replace the legacy Avro code. The current approach is to provide a converter that transforms a DataFrame into a GAMEDataSet, but eventually the GAME and Photon api will consume DataFrame / Dataset directly.

**Note:** This change requires Spark 1.6+, and has been tested on the grid with 1.6.1.
### Performance Comparison

The following documents a performance comparison on an identical proprietary training set before and after this PR.

| Code | AUC | Runtime |
| --- | --- | --- |
| Before | 0.7796065687206302 | 39.55 m |
| After | 0.7801638770709362 | 40.38 m |
### Fixed effect model

| Code | Count | Mean | Variance | L2 norm |
| --- | --- | --- | --- | --- |
| Before | 22468 | 1.985378978052645E-4 | 0.001589526313337537 | 5.976016508476237 |
| After | 22468 | -1.82578704936172E-4 | 0.001650688565248512 | 6.089890718711416 |
### Random effect model 1

| Code | Count | Mean | Std Dev | Max | Min |
| --- | --- | --- | --- | --- | --- |
| Before | 2370903 | 0.317850 | 0.203167 | 3.132743 | 0.000040 |
| After | 2370903 | 0.317734 | 0.203154 | 3.137519 | 0.000037 |
### Random effect model 2

| Code | Count | Mean | Std Dev | Max | Min |
| --- | --- | --- | --- | --- | --- |
| Before | 967254 | 0.136670 | 0.154361 | 2.722700 | 0.000027 |
| After | 967254 | 0.136611 | 0.154208 | 2.719635 | 0.000027 |